Exposing polling functionality for logging.
Summary: Similarly to the usage for frontend polling, not everyone is able to open a websocket-based connection between their logging client and the visdom server. In order to better support this use case, this diff creates a polling version of the reverse socket, which is used with `setup_polling`. It also refactors the final request step out of `_send` to allow more ease in writing visdom wrappers that use alternate request formats.

Reviewed By: lvdmaaten

Differential Revision: D15862043

fbshipit-source-id: 1ec195070670c3b7d84115e08ed832c6a2c44c59