Address issue #399 (#657)
Summary:
## Description

Allows image to resize itself to better fit the window. Resizing is done dynamically and matches its window resize activity. It is also now centered along the larger axis for better presentation.

When the user resizes the window, we want the zoom/drag settings to also reset themselves. This is because the image may go off-frame and disappear. We also handle edge cases where, for example, the container window is taller than it is wide, and the image itself is even taller than the window but has the same width.

## Motivation and Context

This change improves the Visdom user experience.

https://github.com/facebookresearch/visdom/issues/399

## How Has This Been Tested?

Manual test

Rebuild Visdom from modified source code. Run demo.py and add a small 32x32 px image to view changes.

## Screenshots (if appropriate):

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)

## Checklist:

- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] For JavaScript changes, I have re-generated the minified JavaScript code.
Pull Request resolved: https://github.com/facebookresearch/visdom/pull/657

Reviewed By: JackUrb

Differential Revision: D16810898

Pulled By: JackUrb

fbshipit-source-id: 68605ac1b6d72a77fa8d41e2928bd59303eb4e24