Creating 0.1.8.9 release before deprecating py2 and th (#669)
Summary:
## Description
We will be moving to deprecating python 2 support as well as th support. In order to do this, I'm creating a release for the current state of visdom before  handling the deprecation.

## Motivation and Context
Maintaining both the th components of the code as well as dual py2 py3 support is overhead that the visdom project can't handle. We'd like to produce more substantial testing and refactor some older things, and removing these constraints from the design makes moving forward easier. Python 2 is officially being sunset, so it seems like an appropriate time.
Pull Request resolved: https://github.com/facebookresearch/visdom/pull/669

Differential Revision: D17352278

Pulled By: JackUrb

fbshipit-source-id: 0c13a23549d384c45122aa1ef23a30526af752ab