Removing lua torch support (#673)
Summary:
We're removing lua torch support, as it is heavily underused and keeping support for it has demotivated a lot of refactoring in the past. With it gone it should be easier to clean up the codebase.

## Description
This PR removes all of the python code that supports the lua code. I left the lua code present so that people loading lua from the most up-to-date repo will still get warnings properly.

## Motivation and Context
Lua support wasn't being used by many, and the upkeep for keeping the support was making it somewhat difficult to develop new features. As such, `v0.1.8.4` was marked as the last version with lua support, and this PR removes the related code. The `th` folder is left in tact with warnings for users to ensure that people running on the most recent version aren't just left wondering where torch support went.

## How Has This Been Tested?
Ran `demo.py`, everything from the python side worked as expected

## Screenshots (if appropriate):

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing functionality to not work as expected)

## Checklist:

- [x] My code follows the code style of this project.
- [x] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.
- [ ] For JavaScript changes, I have re-generated the minified JavaScript code.
Pull Request resolved: https://github.com/facebookresearch/visdom/pull/673

Reviewed By: pringshia

Differential Revision: D17528864

Pulled By: JackUrb

fbshipit-source-id: 555b97e1c27877281cd9d7494b3a089c90b475dd