Adding support for image click callbacks (#647)
Summary:
Adding support for callbacks on image pane. The coordinates in the image coordinate frame are passed as event properties.

## Description
A new `onClick` event is registered for `ImagePane`. I reuse the existing code in `handleMouseOver` to compute the mouse cursor location in the image coordinate frame. They can be accessed from python as `event['image_coord']['x'], event['image_coord']['y']`. I extended the python demo to register callback with one of the images; click coordinates are logged in a text box; see the screenshot below. As far as I understand, Lua API does not support callbacks at the moment.
I will update the documentation if the change itself looks good.

## Motivation and Context
It is convenient to produce visualisations based on specified anchor points in the image; I use it for my current project.

## How Has This Been Tested?
Running `demo.py`; the new functionality is working (just need to wait until everything is loaded).

## Screenshots (if appropriate):
![Screen Shot 2019-07-19 at 4 55 45 PM](https://user-images.githubusercontent.com/1558328/61550409-4a28cb00-aa4a-11e9-8903-b4c843ef1519.png)

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)

## Checklist:

- [x] My code follows the code style of this project.
- [?] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.
- [x] For JavaScript changes, I have re-generated the minified JavaScript code.
Pull Request resolved: https://github.com/facebookresearch/visdom/pull/647

Differential Revision: D16462288

Pulled By: JackUrb

fbshipit-source-id: 96a87b6bbb5a0c27fa7c9dcea50f7c9de312ac79