Marker border line width override (#623)
Provide a way to change the width of the black border around every point in a
scatter plot from 0.5. A new opt `opts.markerborderwidth` provides a new value.

If you want to use color in a meaningful way in a scatter plot, the
hardcoded black border gets in the way. This is worse when you have
a lot of points in a scatter plot. One use case is plotting an RGB
pointcloud of a scene - every point has a meaningful color, and we
don't want to see black.

- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)

- [X] My code follows the code style of this project.
- [X] My change requires a change to the documentation.
- [X] I have updated the documentation accordingly.
- [ ] For JavaScript changes, I have re-generated the minified JavaScript code.