Marker border line width override (#623) (#696)
Summary:
## Description
Provide a way to change the width of the black border around every point in a
scatter plot from 0.5. A new opt `opts.markerborderwidth` provides the new value.

## Motivation and Context
If you want to use color in a meaningful way in a scatter plot, the
hardcoded black border gets in the way. This is worse when you have
a lot of points in a scatter plot. One use case is plotting an RGB
pointcloud of a scene - every point has a meaningful color, and we
don't want to see black.

I think this is the minimum change to fix the following issue.
https://github.com/facebookresearch/visdom/issues/623

## How Has This Been Tested?
Running on Linux, I observe the difference in safari.
https://snipboard.io/fkPFgi.jpg shows two scatterplots in the demo, one of which has no border. The same works for 3d plots.

## Types of changes
- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)

## Checklist:
- [X] My code follows the code style of this project.
- [X] My change requires a change to the documentation.
- [X] I have updated the documentation accordingly.
- [ ] For JavaScript changes, I have re-generated the minified JavaScript code.
Pull Request resolved: https://github.com/facebookresearch/visdom/pull/696

Differential Revision: D19588227

Pulled By: JackUrb

fbshipit-source-id: 8493c12d27c8e807c855d4b69dfa192de143dc79